requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
python-dotenv==1.0.1
apscheduler==3.10.4
